Refactor backend startup entrypoint
