+++ backend/package.json
@@ -19,7 +19,7 @@
   "scripts": {
     "test": "jest",
     "start": "node startup.js",
-    "start:dev": "node server.js",
+    "start:dev": "node src/server.js",
     "payouts:run": "node backend/tasks/generatePayouts.js",
     "fix-sequences": "node fixSequence.js",
     "prepush": "npm test"
